% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_site_report_table_functions.R
\name{dm_site_tbl}
\alias{dm_site_tbl}
\title{Title}
\usage{
dm_site_tbl(
  population_from,
  population_where = NULL,
  idvar = "USUBJID",
  treatment_var = "TRT01P",
  treatment_order = NULL,
  site_var = "SITEID",
  country_var = NULL,
  level_col_name = FALSE,
  display_total_list = NULL,
  topleft_label = c("", "Site"),
  subgroup_var = NULL,
  subgroup_order = NULL,
  subgroup_label = NULL,
  subgroup_by_col = FALSE,
  rtf = TRUE,
  col_rel_width = NULL,
  column_order = NULL,
  column_group = NULL,
  column_grouped_name = NULL,
  tlf_number = NULL,
  title_text = NULL,
  subtitle_text = NULL,
  studyID = NULL,
  page_header = NULL,
  page_note = NULL,
  page_nrow = 30,
  title_note = NULL,
  foot_note = NULL,
  program_loc = NULL,
  output_file = NULL
)
}
\arguments{
\item{population_from}{Subject-level input dataset, typically ADSL.}

\item{population_where}{A character vector of the filtering condition applied to the subject-level dataset.}

\item{idvar}{A string representing the variable name for subject ID.}

\item{treatment_var}{A string representing the variable name for the treatment.}

\item{treatment_order}{A character vector of treatment groups in the desired order.}

\item{site_var}{A string representing the variable name for the site, default is 'SITEID'.}

\item{country_var}{A string representing the variable name for the country, default is NULL.
This is optional; if not provided, country information will be excluded.}

\item{level_col_name}{A logical parameter indicating whether to include the leveling columns.}

\item{display_total_list}{A list of named lists for aggregating treatment groups.}

\item{topleft_label}{A character vector of top-left column headers.}

\item{subgroup_var}{A string representing the variable name for subject ID.}

\item{subgroup_order}{A character vector of sub groups in the desired order.}

\item{subgroup_label}{A string representing the label for sub groups.}

\item{subgroup_by_col}{A logical parameter indicating whether to include subgroups on same page.}

\item{rtf}{A logical parameter indicating whether to write an RTF file.}

\item{col_rel_width}{A numeric vector of column widths.}

\item{column_order}{A character vector of column order.}

\item{column_group}{A character vector of grouped column names on first row.}

\item{column_grouped_name}{A character vector of customized column names on second row.}

\item{tlf_number}{A string representing the number of table/listing/figure.}

\item{title_text}{A string representing the title.}

\item{subtitle_text}{A string representing the sub title.}

\item{studyID}{A string representing the study ID.}

\item{page_header}{A string representing the page header.}

\item{page_note}{A string representing the page note.}

\item{page_nrow}{A numeric representing the number of rows per page.}

\item{title_note}{A string representing the title note.}

\item{foot_note}{A string representing the foot note.}

\item{program_loc}{A string representing the program location.}

\item{output_file}{A string representing the path to the output file.}
}
\value{
A list containing three data frames: one named report.table,
one named report.table.n for count results,
and the other named report.table.p for percentage results.
}
\description{
Title
}
\examples{
# Define parameters
tlf_number <- "Table 1.2"
title_text <- "Number of Participants Randomized by Site"
subtitle_text <- "Randomized Set"
studyID <- "<Study ID>"
page_header <- "<Page Header>"
page_note <- "<Page Note>"
page_nrow <- 30
title_note <- "<Title Note>"
foot_note <- "<Foot Note>"

# Call the function
dm_site_tbl(
  population_from = toy::demo_adsl,
  country_var = "COUNTRY",
  topleft_label = c("Country", "Site"),
  # subgroup_var = "SEX",
  # subgroup_by_col = TRUE,
  rtf = FASLE,
  tlf_number = tlf_number,
  title_text = title_text,
  subtitle_text = subtitle_text,
  studyID = studyID,
  page_header = page_header,
  page_note = page_note,
  page_nrow = page_nrow,
  title_note = title_note,
  foot_note = foot_note
)
}
